+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### What's Improved
 - Added ability to edit previously played or currently playing items in the queue
 - Added support for date values formatted as "YYYYMMDD"
+- Pressing the button will now clear the current selection before navigating back
 
 #### What's Fixed
 - Fixed unreliable ReplayGain adjustment application in certain situations